pyproject.toml
[tool.poetry]
name = "motor-dados"
version = "0.1.0"
description = "Pipeline simples de ingestão via RabbitMQ e Supabase"
authors = ["Seu Nome <seu.email@exemplo.com>"]


[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pika = "^1.3.2"  # Cliente RabbitMQ
supabase = "^2.13.0"  # Cliente Supabase
python-dotenv = "^1.0.1"  # Carregar variáveis de ambiente
faker = "^36.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"  # Para testes automatizados

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
